% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{bivalent_equation_generator}
\alias{bivalent_equation_generator}
\title{bivalent_equation_generator
Generate equation according to ploidy level under assumed bivalent pairing}
\usage{
bivalent_equation_generator(ploidy)
}
\arguments{
\item{ploidy}{crop ploidy level}
}
\value{
a list include two formats of results (probability, recombination) and include
format 1 is contingency table. format 2 is melt table
}
\description{
bivalent_equation_generator
Generate equation according to ploidy level under assumed bivalent pairing
}
