% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{check_segregation}
\alias{check_segregation}
\title{examine segregation inheritance}
\usage{
check_segregation(
  possible_gamete = P1_gamete_diso,
  parent = "P1",
  probability_offspring = F1,
  seg_invalidrate,
  missing_nbr = missing_nbr
)
}
\arguments{
\item{possible_gamete}{expected possible gametes of one parent. It is generated either from
*create_gamete_disomic* or *create_gamete_polysomic*}

\item{parent}{'P1' or 'P2'}

\item{probability_offspring}{probability offspring scores which are used for test.It is a dataframe contains:
+ indivdual: individual name
+ P1_1, P1_2: the gametes from P1 (not phased)
+ P2_1, P2_2: the gametes from P2 (not phased)
+ probability: the probability obtaining the gametes
+ pairing: the pairing of two parents
+ pair_freq: the frequency of this pairing from two parents
+ P1_gamete_freq, P2_gamete_freq: Under this assumed pairing and inheritance, what is the expected
segregation ratio of obtaining this gamete.}

\item{seg_invalidrate}{the value used in binomial test}

\item{missing_nbr}{the number of missing individuals}
}
\value{

}
\description{
examine segregation inheritance
}
