% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{choose_offspringscore}
\alias{choose_offspringscore}
\title{based on tested pairing to choose to offspring score}
\usage{
choose_offspringscore(offspring_score_list, pairing)
}
\arguments{
\item{offspring_score_list}{It includes two lists: input and output,
The 'input' include: P1_inheritance, P2_inheritance, ploidy. The output include a list of all markers. Within each marker,
it include the missing_nbr(estimated number of missing individuals), and all possible pairing configuration, what is the probability
of each gametes. It is a dataframe contains:
+ indivdual: individual name
+ P1_1, P1_2: the gametes from P1 (not phased)
+ P2_1, P2_2: the gametes from P2 (not phased)
+ probability: the probability obtaining the gametes
+ pairing: the pairing of two parents
+ pair_freq: the frequency of this pairing from two parents
+ P1_gamete_freq, P2_gamete_freq: Under this assumed pairing and inheritance, what is the expected
segregation ratio of obtaining this gamete.}

\item{pairing}{a list include
+ output: a data frame include the marker name and the pairing each marker follows of the target parent
+ target_parent: which parent follows disomic inheritance
+ inheritance scores: to determine the pairing configuration, mode of inheritance study was performed,
what is the inheritance scores and parameters in it. This is a combined list include each marker, under each
pairing configuration, their: chi-square value, binomial_p, multiplied_p, combined_p, observed_count,
expected count, and invalid percentage.}
}
\value{
It includes two lists: input and output,
The 'input' include: P1_inheritance, P2_inheritance, ploidy. The output include a list of all markers. Within each marker,
it include the missing_nbr(estimated number of missing individuals), and all possible pairing configuration, what is the probability
of each gametes. It is a dataframe contains:
+ indivdual: individual name
+ P1_1, P1_2: the gametes from P1 (not phased)
+ P2_1, P2_2: the gametes from P2 (not phased)
+ probability: the probability obtaining the gametes
+ pairing: the pairing of two parents
+ pair_freq: the frequency of this pairing from two parents
+ P1_gamete_freq, P2_gamete_freq: Under this assumed pairing and inheritance, what is the expected
segregation ratio of obtaining this gamete.
Please note here only the chosen pairing dataframe is kept
}
\description{
based on tested pairing to choose to offspring score
}
