% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{create_bivalent_pairing}
\alias{create_bivalent_pairing}
\title{create_bivalent_pairing
generate the format of bivalent pairing according to the ploidy level}
\usage{
create_bivalent_pairing(ploidy)
}
\arguments{
\item{ploidy}{crop ploidy level}
}
\value{
a vector of all possible pairing in two formats
In hexaploid, format1 is: 
 "12|34|56" "12|35|46" "12|36|45" "13|26|45" "13|25|46" "13|24|56" "14|23|56" "16|23|45" "15|23|46"
 format2 is:
 "H1_H2|H3_H4|H5_H6" "H1_H2|H3_H5|H4_H6" "H1_H2|H3_H6|H4_H5" "H1_H3|H2_H6|H4_H5" "H1_H3|H2_H5|H4_H6" "H1_H3|H2_H4|H5_H6"
 "H1_H4|H2_H3|H5_H6" "H1_H6|H2_H3|H4_H5" "H1_H5|H2_H3|H4_H6"
}
\description{
create_bivalent_pairing
generate the format of bivalent pairing according to the ploidy level
}
