% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{create_gamete_disomic}
\alias{create_gamete_disomic}
\title{create_gamete_disomic
generate the disomic inheritance gametes under all situations}
\usage{
create_gamete_disomic(parent = P1, ploidy)
}
\arguments{
\item{parent}{one parent's multi-allelic marker, e.g. AAAC}
}
\value{
all possible gametes under disomic inheritance of all possibility situations
}
\description{
create_gamete_disomic
generate the disomic inheritance gametes under all situations
}
