% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{create_gamete_polysomic}
\alias{create_gamete_polysomic}
\title{create_gamete_polysomic
generate the polysomic inheritance gametes}
\usage{
create_gamete_polysomic(parent = P1, ploidy)
}
\arguments{
\item{parent}{one parent's multi-allelic marker, e.g. AAAC}
}
\value{
all possible gametes under polysomic inheritance
}
\description{
create_gamete_polysomic
generate the polysomic inheritance gametes
}
