% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{dataframe_to_contigency}
\alias{dataframe_to_contigency}
\title{dataframe_to_contigency
convert dataframe to contingency table}
\usage{
dataframe_to_contigency(tmp, col1 = "A", col2 = "B", count = "count")
}
\arguments{
\item{tmp}{a dataframe}

\item{col1}{the colname which will be converted in contingency table as row}

\item{col2}{the colname which will be converted in contingency table as col}

\item{count}{the colname which will be used to fill in the contingency table}
}
\value{
contigency table after required conversion
}
\description{
dataframe_to_contigency
convert dataframe to contingency table
}
