% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{decide_MT}
\alias{decide_MT}
\title{decide_MT
marker type identification}
\usage{
decide_MT(parent = P1)
}
\arguments{
\item{parent}{parent genotype}
}
\value{
turn the parent genotype of multi-allelic marker to a number. 
ABCD: 1_1_1_1
AABC: 1_1_2
AABB: 2_2
AAAB: 1_3
AAAA: 4
}
\description{
decide_MT
marker type identification
}
