% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{disomic_equation_generator}
\alias{disomic_equation_generator}
\title{generate equation according to ploidy level under assumed disomic inheritance}
\usage{
disomic_equation_generator(ploidy)
}
\arguments{
\item{ploidy}{crop ploidy level}
}
\value{
a list include two formats of results (probability, recombintion) and include
format 1 is contigency table. format 2 is melt table. Each format is a list include all po
possible pairing's result.
}
\description{
generate equation according to ploidy level under assumed disomic inheritance
}
