% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{distance_to_rf}
\alias{distance_to_rf}
\title{distance_to_rf
Convert chromosomal position to difference of recombination frequency}
\usage{
distance_to_rf(dt)
}
\arguments{
\item{dt}{integer for the distance of locus 1 and locus 2}
}
\value{
a number of converted recombination frequency
}
\description{
distance_to_rf
Convert chromosomal position to difference of recombination frequency
}
