% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{estimated_phasing}
\alias{estimated_phasing}
\title{estimated_phasing
Extract estimated genotype phasing for each marker}
\usage{
estimated_phasing(mrk1_phase, mrk2_phase)
}
\arguments{
\item{mrk1_phase}{string of marker P estimated phasing}

\item{mrk2_phase}{string of marker Q estimated phasing}
}
\value{

}
\description{
estimated_phasing
Extract estimated genotype phasing for each marker
}
