% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{format_pairint_list}
\alias{format_pairint_list}
\title{format_pairint_list}
\usage{
format_pairint_list(temp = poly_total, F1_ind)
}
\arguments{
\item{temp}{a dataframe contains the possible offspring genotype and their corresponding
probability.}

\item{F1_ind}{individual name of all F1 individuals}
}
\value{
a list for each paring, including the locus positions for marker P and Q,
and their probability of recombination, and the count from offspring genotypes.
}
\description{
format_pairint_list
}
