% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{format_results}
\alias{format_results}
\title{format_results
format result from PedigreeSim format to Multi-allelic mode inheritance format}
\usage{
format_results(dosages = poly_dosages, ploidy = 4, Parent1, Parent2, ncores)
}
\arguments{
\item{dosages}{Pedigreesim simulated multi-allelic marker format. Markers as row, individuals as column.}

\item{ploidy}{crop ploidy level}

\item{Parent1}{the column name of parent 1}

\item{Parent2}{the column name of parent 2}

\item{ncores}{number of cores used for running, suggested to use -2 cores of your computers}
}
\value{
formatted list of each haploblock (temp, parent)
}
\description{
format_results
format result from PedigreeSim format to Multi-allelic mode inheritance format
}
