% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{generate_all_equation_polysomic}
\alias{generate_all_equation_polysomic}
\title{generate all possible equation under polysomic inheritance}
\usage{
generate_all_equation_polysomic(
  P_genotype = strsplit("AABBBB", "")[[1]],
  Q_genotype = strsplit("ABCDEF", "")[[1]],
  generalized_equation
)
}
\arguments{
\item{P_genotype}{the genotype of marker P at one parent. It is a vector of characters, e.g. c("A","A","B","B")}

\item{Q_genotype}{the genotype of marker Q also at this parent. It is a vector of characters}

\item{generalized_equation}{generalized equation obtained from *disomic_equation_generator* according to the ploidy level}
}
\value{
a list include when parents with the given genotype, under each possible phasing, what is the 
expected probability and number of recombination
}
\description{
generate all possible equation under polysomic inheritance
}
