% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{generate_recombi_nonrecombi}
\alias{generate_recombi_nonrecombi}
\title{generate_recombi_nonrecombi
Generate the possible allele combinations together with their possibilities}
\usage{
generate_recombi_nonrecombi(
  hom_tmp = tmp_P[, c(part1_hom1, part1_hom2)],
  ploidy
)
}
\arguments{
\item{hom_tmp}{a matrix of allele positions with their chromosomes, as the format of:
     H1  H2 
[1,] "1" "2"
[2,] "1" "2"}

\item{ploidy}{crop ploidy level}
}
\value{
a list include two sublist with the recombination and non recombination 
happened probability, as the format of:
$`1/2*s`
[1] "11" "22"

$`1/2*r`
[1] "12" "21"
}
\description{
generate_recombi_nonrecombi
Generate the possible allele combinations together with their possibilities
}
