% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{Haplotype_linkage}
\alias{Haplotype_linkage}
\title{perform linkage anlaysis in chosen marker type}
\usage{
Haplotype_linkage(
  Mrk_P = combs[c, ][1],
  Mrk_Q = combs[c, ][2],
  offspring_score_list = offspring_score_list1,
  poly,
  cores,
  all_equations,
  pairing
)
}
\arguments{
\item{Mrk_P}{one marker, it need to be in the format of "AABB_EEFF"}

\item{Mrk_Q}{another marker, it need to be in the format of "AABB_EEFF"}

\item{offspring_score_list}{It includes two lists: input and output,
The 'input' include: P1_inheritance, P2_inheritance, ploidy. The output include a list of all markers. Within each marker,
it include the missing_nbr(estimated number of missing individuals), and all possible pairing configuration, what is the probability
of each gametes. It is a dataframe contains:
+ indivdual: individual name
+ P1_1, P1_2: the gametes from P1 (not phased)
+ P2_1, P2_2: the gametes from P2 (not phased)
+ probability: the probability obtaining the gametes
+ pairing: the pairing of two parents
+ pair_freq: the frequency of this pairing from two parents
+ P1_gamete_freq, P2_gamete_freq: Under this assumed pairing and inheritance, what is the expected
segregation ratio of obtaining this gamete.}

\item{poly}{formatted list include all individuals and parents genotype. It is a list of all marker's names
Within each marker, it includes:
+ parental info: a dataframe include two parents' genotype
+ temp: a dataframe include alll individuals' genotype
All the information here is not phased.}

\item{cores}{number of cores used to run the analysis. It is suggested to use less than your computer capacity. To
check how many cores you computer has, you can use *detectCores()*}

\item{all_equations}{a list include:
- equations: equations for P1 and P2 separately
- input: P1 and P2's inheritance, ploidy level, and marker type (MT)}

\item{pairing}{a list include
+ output: a data frame include the marker name and the pairing each marker follows of the target parent
+ target_parent: which parent follows disomic inheritance
+ inheritance scores: to determine the pairing configuration, mode of inheritance study was performed,
what is the inheritance scores and parameters in it. This is a combined list include each marker, under each
pairing configuration, their: chi-square value, binomial_p, multiplied_p, combined_p, observed_count,
expected count, and invalid percentage.}
}
\value{
a list include input and output. In the input, it has: ploidy, Mrk_P, and Mrk_Q.
In the output, it includes the linkage estimation result
}
\description{
perform linkage anlaysis in chosen marker type
}
