% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{Infer_offspring_inheritance}
\alias{Infer_offspring_inheritance}
\title{Infer offspring inheritance}
\usage{
Infer_offspring_inheritance(poly, ploidy, P1_inheritance, P2_inheritance)
}
\arguments{
\item{poly}{formatted list include all individuals and parents genotype. It is a list of all marker's names
Within each marker, it includes:
+ parental info: a dataframe include two parents' genotype
+ temp: a dataframe include alll individuals' genotype
All the information here is not phased.}

\item{ploidy}{crop ploidy level}

\item{P1_inheritance}{either 'Polysomic' or 'Disomic'}

\item{P2_inheritance}{either 'Polysomic' or 'Disomic'}
}
\value{
It includes two lists: input and output,
The 'input' include: P1_inheritance, P2_inheritance, ploidy. The output include a list of all markers. Within each marker,
it include the missing_nbr(estimated number of missing individuals), and all possible pairing configuration, what is the probability
of each gametes. It is a dataframe contains:
+ indivdual: individual name
+ P1_1, P1_2: the gametes from P1 (not phased)
+ P2_1, P2_2: the gametes from P2 (not phased)
+ probability: the probability obtaining the gametes
+ pairing: the pairing of two parents
+ pair_freq: the frequency of this pairing from two parents
+ P1_gamete_freq, P2_gamete_freq: Under this assumed pairing and inheritance, what is the expected
segregation ratio of obtaining this gamete.
}
\description{
Infer offspring inheritance
}
