% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{MarkerType_Overview}
\alias{MarkerType_Overview}
\title{Have an overview on marker type and the sharing of haplotypes between two parents}
\usage{
MarkerType_Overview(poly, ploidy, plot = TRUE)
}
\arguments{
\item{poly}{}

\item{ploidy}{crop ploidy level}

\item{plot}{it can be either True/False. If true, the overview graph will be generated}
}
\value{
a dataframe record each markers' marker type of two parents (MT_P1, MT_P2),the
numer of shared haplotypes between two parents, and the unique haplotype
}
\description{
Have an overview on marker type and the sharing of haplotypes between two parents
}
