% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{maximum_likelihood_estimator}
\alias{maximum_likelihood_estimator}
\title{use maximumlikelihood method to estimate the rf}
\usage{
maximum_likelihood_estimator(equation = maximum_equation, plot)
}
\arguments{
\item{equation}{maximumlikelihood function. Here is a string}

\item{plot}{if TRUE, how the function looks like will be plotted}
}
\value{
a list containing: 'rf'
}
\description{
use maximumlikelihood method to estimate the rf
}
