% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{Multi_allelic_simulator_rep}
\alias{Multi_allelic_simulator_rep}
\title{Simulate multi_allelic markers by using PedigreeSim
the marker can be specified with different types. At each locus, it will include all marker types}
\usage{
Multi_allelic_simulator_rep(
  chrnum = 1,
  gap = 1,
  length = 99,
  ploidy = 4,
  filename = "test",
  popnum = 200,
  marker_type = pair_1,
  pairing = 0,
  quadrivalents = 0
)
}
\arguments{
\item{chrnum}{numeric. Number of chromosomes}

\item{gap}{numeric. the gap length}

\item{length}{numeric. Total length of chromosome}

\item{ploidy}{numeric. crop ploidy level}

\item{filename}{string. filename which will be used to save files}

\item{popnum}{numeric. The number of F1 individuals}

\item{marker_type}{a vector of strings in the format of c("ABCD_EFGH","AABC_EEFG","AAAB_EEEF")}

\item{pairing}{numeric. 0 - polysomic; 1 - disomic}

\item{quadrivalents}{numeric. 0 - bivalent pairing; 1 - multivalent pairing}
}
\value{

}
\description{
Simulate multi_allelic markers by using PedigreeSim
the marker can be specified with different types. At each locus, it will include all marker types
}
