% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{MultiAllelic_Inheritance_simulator}
\alias{MultiAllelic_Inheritance_simulator}
\title{Simulate polysomic x disomic dataset with multi-allelic markers}
\usage{
MultiAllelic_Inheritance_simulator(
  filename = "Pedtry",
  marker_type = c("ABCD_EFGH", "AABB_EEFF"),
  pref_P1 = 0,
  quad_P1 = 1,
  pref_P2 = 0.5,
  quad_P2 = 0,
  ploidy = 4,
  popnum = 200,
  lg = 1,
  gap = 1,
  length = 99
)
}
\arguments{
\item{filename}{string. filename which will be used to save files}

\item{marker_type}{a vector of strings in the format of c("ABCD_EFGH","AABC_EEFG","AAAB_EEEF")}

\item{pref_P1}{numeric. 0 - polysomic; 1 - disomic (for parent 1)}

\item{quad_P1}{numeric. 0 - bivalent pairing; 1 - multivalent pairing (for parent 1)}

\item{pref_P2}{numeric. 0 - polysomic; 1 - disomic (for parent 2)}

\item{quad_P2}{numeric. 0 - bivalent pairing; 1 - multivalent pairing (for parent 2)}

\item{ploidy}{numeric. crop ploidy level}

\item{popnum}{numeric. The number of F1 individuals}

\item{lg}{numeric. Number of chromosomes}

\item{gap}{numeric. the gap length}

\item{length}{numeric. Total length of chromosome}
}
\value{
different types of files with prefix of your specified filename
}
\description{
Simulate polysomic x disomic dataset with multi-allelic markers
}
