% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{multiplot}
\alias{multiplot}
\title{multiplot
Generating plots in one display window}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{name of plots want to combine}

\item{cols}{number of plots want to display in columns}

\item{layout}{number of plots want to display in rows}
}
\value{
arranged plot displays
}
\description{
multiplot
Generating plots in one display window
}
