% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{offspring_probability_estimator}
\alias{offspring_probability_estimator}
\title{offspring_probability_estimator
generate the probability for offspring's genotype under specified inheritance}
\usage{
offspring_probability_estimator(
  F1 = offspring_score,
  P1,
  P2,
  P1_inheritance,
  P2_inheritance,
  ploidy = 4
)
}
\arguments{
\item{F1}{dataframe of offspring's multi-allelic marker genotype}

\item{P1}{parent1's genotype}

\item{P2}{parent2's genotype}

\item{P1_inheritance}{inheritance of P1, can be either 'Polysomic' or 'Disomic'}

\item{P2_inheritance}{inheritance of P2, can be either 'Polysomic' or 'Disomic'}

\item{ploidy}{crop ploidy level}
}
\value{
a dataframe include each offspring's genotype probability under different mode of inheritance
}
\description{
offspring_probability_estimator
generate the probability for offspring's genotype under specified inheritance
}
