% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{pairing_examinator}
\alias{pairing_examinator}
\title{examine which pairing configuration the parent follow disomic inheritance goes
This function aims to help the parents who follows disomic inheritance to decide whith pairing
it follows exactly. This choice is made by using the F1 segregation ratio.}
\usage{
pairing_examinator(
  MT = MT,
  seg_invalidrate = 0.05,
  offspring_score_list = offspring_score_list
)
}
\arguments{
\item{MT}{a dataframe generated from *MarkerType_Overview*. It needs to include the columns:
marker (marker name), parents (parents genotype), P1_MT (P1 marker type), P2_MT (P2_marker type),
sharing (the number of shared haplotypes between two parents), unique_hap, MT (combined marker type)}

\item{seg_invalidrate}{the value used in binomial test}

\item{offspring_score_list}{a list generated from *Infer_offspring_inheritance*. It needs to include two lists: input and output,
The 'input' include: P1_inheritance, P2_inheritance, ploidy. The output include a list of all markers. Within each marker,
it include the missing_nbr(estimated number of missing individuals), and all possible pairing configuration, what is the probability
of each gametes. It is a dataframe contains:
+ indivdual: individual name
+ P1_1, P1_2: the gametes from P1 (not phased)
+ P2_1, P2_2: the gametes from P2 (not phased)
+ probability: the probability obtaining the gametes
+ pairing: the pairing of two parents
+ pair_freq: the frequency of this pairing from two parents
+ P1_gamete_freq, P2_gamete_freq: Under this assumed pairing and inheritance, what is the expected
segregation ratio of obtaining this gamete.}
}
\value{
a list include
+ output: a data frame include the marker name and the pairing each marker follows of the target parent
+ target_parent: which parent follows disomic inheritance
+ inheritance scores: to determine the pairing configuration, mode of inheritance study was performed,
what is the inheritance scores and parameters in it. This is a combined list include each marker, under each
pairing configuration, their: chi-square value, binomial_p, multiplied_p, combined_p, observed_count,
expected count, and invalid percentage.
}
\description{
examine which pairing configuration the parent follow disomic inheritance goes
This function aims to help the parents who follows disomic inheritance to decide whith pairing
it follows exactly. This choice is made by using the F1 segregation ratio.
}
