% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{plot_rf_comparison}
\alias{plot_rf_comparison}
\title{plot evaluated results}
\usage{
plot_rf_comparison(result, filename)
}
\arguments{
\item{result}{a dataframe include the linkage results and simulated r and phasing}

\item{filename}{saved filename}
}
\value{

}
\description{
plot evaluated results
}
