% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{possible_gamete_parent}
\alias{possible_gamete_parent}
\title{find possible gamete of parent}
\usage{
possible_gamete_parent(tmp, loc = "P", ploidy)
}
\arguments{
\item{tmp}{a dataframe has two rows represent the genotype of each parent}

\item{loc}{a character, i.e. 'P' which is one of the rownames of *tmp*}

\item{ploidy}{crop ploidy level}
}
\value{
a vector including all possible gametes that can produced from the
specified parent
}
\description{
find possible gamete of parent
}
