% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{readDatfile}
\alias{readDatfile}
\title{readDatfile
Importing the data with R function read.table.}
\usage{
readDatfile(file, header = TRUE, sep = "\\t", check.names = FALSE, ...)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{sep}{the field separator character.}

\item{check.names}{logical. If TRUE then the names of the variables in the data frame are checked to 
ensure that they are syntactically valid variable names.}
}
\value{
a matrix contain the dosage information for all individuals, 
according to their markers and alleles.
}
\description{
readDatfile
Importing the data with R function read.table.
}
