% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{recombination_estimator}
\alias{recombination_estimator}
\title{estimate recombination frequency using MML per marker pair}
\usage{
recombination_estimator(
  equation = Q_P2_equation,
  tmp = offspring_probability_table,
  share_nbr,
  pairing,
  mrk = marker_Q,
  parent = "P2"
)
}
\arguments{
\item{equation}{the corresponding equation of the chosen marker}

\item{tmp}{corresponding offspring probability table}

\item{share_nbr}{True/False. When it is True, it share more than one haplotype}

\item{pairing}{a list include
+ output: a data frame include the marker name and the pairing each marker follows of the target parent
+ target_parent: which parent follows disomic inheritance
+ inheritance scores: to determine the pairing configuration, mode of inheritance study was performed,
what is the inheritance scores and parameters in it. This is a combined list include each marker, under each
pairing configuration, their: chi-square value, binomial_p, multiplied_p, combined_p, observed_count,
expected count, and invalid percentage.}

\item{mrk}{the marker name of marker Q at this comparison}

\item{parent}{the target parent, 'P1' or 'P2'}
}
\value{
a dataframe include all phasing situations' estimated rf
}
\description{
estimate recombination frequency using MML per marker pair
}
