% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{seg_pro_NEW}
\alias{seg_pro_NEW}
\title{convert the offspring information to required format}
\usage{
seg_pro_NEW(F1 = offspring_temp, parent = parent)
}
\arguments{
\item{F1}{offspring data from poly. A dataframe with ploidy number's column
and individual numbers row. The genotype within each individual is not phased}

\item{parent}{'P',the prefix of parent}
}
\value{
the counted frequency of offspring genotype
}
\description{
convert the offspring information to required format
}
