% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{sharing_identification}
\alias{sharing_identification}
\title{identify the sharing between two parents}
\usage{
sharing_identification(mrk)
}
\arguments{
\item{mrk}{a string record one marker. e.g. 'AABC_AABE'}
}
\value{
a numeric about the percentage of sharing between two parents
}
\description{
identify the sharing between two parents
}
