% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{simplify_equation}
\alias{simplify_equation}
\title{simplify_equation
simplify the equation according to specific marker type
the general matrix can be folded according to specific marker type}
\usage{
simplify_equation(
  equation = tetraploid_equation,
  comb_P = a_parents["P1", ],
  comb_Q = combs_b_P1[z, ],
  ploidy = 4
)
}
\arguments{
\item{equation}{equation generated according to ploidy and pairing}

\item{comb_P}{a matrix including the genotype of one marker's one parent, e.g. matrix('A','A','B','C')}

\item{comb_Q}{a matrix including the genotype of the other marker's one parent, e.g. matrix('A','A','B','C')}

\item{ploidy}{crop ploidy level}
}
\value{
a dataframe including all
}
\description{
simplify_equation
simplify the equation according to specific marker type
the general matrix can be folded according to specific marker type
}
