% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{simulated_distance}
\alias{simulated_distance}
\title{simulated_distance
Calculate the distance between real simulated marker positions and model estimated ones}
\usage{
simulated_distance(mapfile, mrk1, mrk2)
}
\arguments{
\item{mapfile}{the marker position file for stimulated dataset}

\item{mrk1}{the marker P in linkage estimation result}

\item{mrk2}{the marker Q in linkage estimation result}
}
\value{
integer for the distance of locus 1 and locus 2
}
\description{
simulated_distance
Calculate the distance between real simulated marker positions and model estimated ones
}
