% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{simulated_phasing}
\alias{simulated_phasing}
\title{simulated_phasing
Extract real genotype phasing for each marker}
\usage{
simulated_phasing(genfile, mrk1, mrk2)
}
\arguments{
\item{genfile}{the genotype phasing file for stimulated dataset}

\item{mrk1}{the marker P name in linkage estimation result}

\item{mrk2}{the marker Q name in linkage estimation result}
}
\value{

}
\description{
simulated_phasing
Extract real genotype phasing for each marker
}
