% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiding_functions.R
\name{unique_hap}
\alias{unique_hap}
\title{find the unique haplotype configuration}
\usage{
unique_hap(hap_mrkP = a_parents["P1", ], hap_mrkQ = combs_b_P1)
}
\arguments{
\item{pair_comb}{a dataframe include two columns, first column include P1's haplotype (e.g. ABCD), second column include 
P2's haplotype}
}
\value{
a same dataframe with the unique configuration (filter out the replicate configuration)
}
\description{
find the unique haplotype configuration
}
